split AnimateApp
